transmedialint/requirements.txt
django
django-filter
djangorestframework
psycopg2-binary
urllib3
requests
bs4
html5lib
pytz
IPython
gensim
bokeh
numpy
pandas
spacy
python-scrapyd-api
scrapyd-client
scrapyd
scrapy-splash
html2text
